refacror MainMenuState
