+++ assets/Interface/CreditsControl.xml
@@ -708,7 +708,7 @@
             </panel>
 
             <effect>
-                <onActive name="autoScroll" length="70000" start="0" end="-8500" inherit="true" onEndEffect="restartCredits()" />
+                <onActive name="autoScroll" length="70000" start="0" end="-8500" inherit="true" onEndEffect="$action" />
             </effect>
         </panel>
     </controlDefinition>
